// src/models/CategoryDTO.ts
export interface CategoryDTO {
    categoryId?: string;
    categoryName?: string;
    updatedDate?: string;
    updatedBy?: string;
    type?: string;
    quantity?: number;
    note?: string;
    // artworkId?: string;
    // artworkName?: string;
    // price?: number;
    // discount?: number;
    // status?: string;
    // id?: string;
    // categoryId?: string;
    // imageUrl?: string;
    // fPosts?: [];
    // fReports?: []
}